cpe var does not keep any status anymore
